# src/routes/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Darve Server API",
    "description": "API documentation for the Darve Server including users, posts, notifications, auth, discussions, tasks, wallet, and follows",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "",
      "description": "Development server"
    }
  ],
  "security": [],
  "paths": {
...
    "/api/discussions": {
      "get": {
        "tags": ["Discussions"],
        "summary": "Get discussions",
        "description": "Get discussions for the current user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of discussions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Discussion"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Discussions"],
        "summary": "Create discussion",
        "description": "Create a new discussion",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiscussion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Discussion created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discussion"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/discussions/{discussion_id}": {
      "patch": {
        "tags": ["Discussions"],
        "summary": "Update discussion",
        "description": "Update a discussion",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "description": "ID of the discussion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiscussion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Discussion updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": ["Discussions"],
        "summary": "Delete discussion",
        "description": "Delete a discussion",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "description": "ID of the discussion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discussion deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/discussions/{discussion_id}/tasks": {
      "post": {
        "tags": ["Discussions"],
        "summary": "Create task for discussion",
        "description": "Create a task related to a discussion",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "description": "ID of the discussion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequestInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": ["Discussions"],
        "summary": "Get tasks for discussion",
        "description": "Get tasks related to a discussion",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "description": "ID of the discussion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks for the discussion",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRequestView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/discussions/{discussion_id}/chat_users": {
      "post": {
        "tags": ["Discussions"],
        "summary": "Add users to discussion",
        "description": "Add users to a discussion chat",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "description": "ID of the discussion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscussionUsers"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/DiscussionUsers"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Users added to discussion"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": ["Discussions"],
        "summary": "Remove users from discussion",
        "description": "Remove users from a discussion chat",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "description": "ID of the discussion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscussionUsers"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/DiscussionUsers"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Users removed from discussion"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/discussions/{discussion_id}/sse": {
      "get": {
        "tags": ["Discussions"],
        "summary": "Discussion SSE stream",
        "description": "Server-sent events stream for discussion updates",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "description": "ID of the discussion",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topic_id",
            "in": "query",
            "description": "Filter by topic ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of discussion updates",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/discussions/{discussion_id}/posts": {
      "post": {
        "tags": ["Discussions"],
        "summary": "Create post in discussion",
        "description": "Create a new post in a discussion",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "description": "ID of the discussion",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PostInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TaskRequestView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Task request ID"
          },
          "created_by": {
            "$ref": "#/components/schemas/UserView"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskRequestViewParticipant"
            },
            "description": "Task participants"
          },
          "request_txt": {
            "type": "string",
            "description": "Task request description"
          },
          "donors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskRequestDonorView"
            },
            "description": "Task donors"
          },
          "reward_type": {
            "$ref": "#/components/schemas/RewardType"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencySymbol"
          },
          "wallet_id": {
            "type": "string",
            "description": "Wallet ID"
          }
        }
      },
      "TaskRequestViewParticipant": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserView"
          },
          "status": {
            "$ref": "#/components/schemas/TaskParticipantStatus"
          }
        }
      },
      "TaskRequestDonorView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Donor ID"
          },
          "user": {
            "$ref": "#/components/schemas/UserView"
          },
          "amount": {
            "type": "integer",
            "format": "int64",
            "description": "Donation amount"
          }
        }
      },
      "UserView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "full_name": {
            "type": "string",
            "description": "Full name"
          }
        }
      },
      "TaskParticipantStatus": {
        "type": "string",
        "enum": ["Requested", "Rejected", "Accepted", "Delivered"],
        "description": "Status of task participant"
      },
      "RewardType": {
        "type": "string",
        "description": "Type of reward for the task"
      },
      "CurrencySymbol": {
        "type": "string",
        "description": "Currency symbol"
      },
...
      "PostInput": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 5,
            "description": "Post title (minimum 5 characters)"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "description": "Post content (minimum 1 characters)"
          },
          "topic_id": {
            "type": "string",
            "description": "Topic ID for the post"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "description": "Tags for the post (maximum 5)"
          },
          "file_1": {
            "type": "string",
            "format": "binary",
            "description": "Optional file attachment"
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Post ID"
          },
          "belongs_to": {
            "type": "string",
            "description": "Discussion ID this post belongs to"
          },
          "created_by": {
            "type": "string",
            "description": "User ID who created the post"
          },
          "title": {
            "type": "string",
            "description": "Post title"
          },
          "content": {
            "type": "string",
            "description": "Post content"
          },
          "media_links": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Media links attached to the post"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags associated with the post"
          },
          "likes_nr": {
            "type": "integer",
            "description": "Number of likes"
          },
          "replies_nr": {
            "type": "integer",
            "description": "Number of replies"
          },
          "r_created": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "r_updated": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "Discussion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Discussion ID"
          },
          "title": {
            "type": "string",
            "description": "Discussion title"
          },
          "description": {
            "type": "string",
            "description": "Discussion description"
          },
          "created_by": {
            "type": "string",
            "description": "User ID who created the discussion"
          },
          "private_discussion_user_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User IDs for private discussions"
          },
          "r_created": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          }
        }
      },
      "CreateDiscussion": {
        "type": "object",
        "required": ["community_id", "title", "private_discussion_users_final"],
        "properties": {
          "community_id": {
            "type": "string",
            "example": "community-123"
          },
          "title": {
            "type": "string",
            "minLength": 5,
            "example": "My discussion title"
          },
          "image_uri": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "example": "https://example.com/image.png"
          },
          "chat_user_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "example": ["user1", "user2"]
          },
          "private_discussion_users_final": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "UpdateDiscussion": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Discussion title"
          },
          "description": {
            "type": "string",
            "description": "Discussion description"
          }
        }
      },
      "DiscussionUsers": {
        "type": "object",
        "required": ["user_ids"],
        "properties": {
          "user_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of user IDs"
          }
        }
      },
      "TaskRequestInput": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": {
            "type": "string",
            "minLength": 5,
            "description": "Task request description (min 5 characters)"
          },
          "participant": {
            "type": "string",
            "nullable": true,
            "description": "Optional participant identifier"
          },
          "offer_amount": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "nullable": true,
            "description": "Optional offer amount, must be at least 1"
          },
          "acceptance_period": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "nullable": true,
            "description": "Optional acceptance period in days"
          },
          "delivery_period": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "nullable": true,
            "description": "Optional delivery period in days"
          }
        }
      },
      "TaskRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Task request ID"
          },
          "created_by": {
            "type": "string",
            "description": "User ID who created the task"
          },
          "request_txt": {
            "type": "string",
            "description": "Task request description"
          },
          "reward_type": {
            "$ref": "#/components/schemas/RewardType"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencySymbol"
          },
          "wallet_id": {
            "type": "string",
            "description": "Wallet ID"
          },
          "r_created": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    }
  },
...
}